Reduce the number of temporary tensors when loading model parameters.
Let the loop iterate over the parameters of the model rather than
over the tensors in the files. This is more efficient because it
loads the tensor from the file only when it is need to be filled into
the model, or in other words, the files are read on demand.