Use internal channels for backend change synchronization.
The `pie` CLI previously used a client to wait until the backend was
attached. However, the CLI and server components are both internal to
`pie`, so letting the CLI part create a client which in turn connects to
the server through a public websocket complicates the code.

This commit refactors the synchronization between the CLI and server by
using internal channels between them.