Fix a race condition in the runtime when launching inferlets.
The runtime stores the join handle in a map after starting an instance.
However, the started instance may trigger an event that requires the
join handle to be in the map.

This commit fixes the race condition by starting the instance after the
join handle is stored into the map.