Use internal authentication for Pie shell and backend connections.
Previously, the user CLI client, the Pie shell, and the backend
subprocess all used the same authentication token for authentication.
However, since the Pie shell and backend subprocesses are part of the
Pie system, they should use a simplified authentication mechanism.

For user CLI client, the authentication token will later be changed to
public key authentication.

This commit adds a new internal authentication mechanism for the Pie
shell and backend subprocesses. The internal authentication token is a
random 64-character string, which is shared with the Pie shell and the
backend subprocesses internally.