Fix premature send pump termination on session drop.
Previously, when a session was dropped, the send pump was terminated
using `abort()`, which would cause any pending messages to be dropped.

This has caused an error that when client authentication failed, the
client would not receive the authentication failure message.

This commit fixes this by not aborting the send pump when the session is
dropped, instead letting the `ws_msg_tx` object be dropped, which will
cause the send pump to terminate after it has sent all pending messages.