Support automatic key detection in `pie-cli config init`.
Automatically detect the private key in `~/.ssh` and use it if found.
Search in the order of `id_ed25519`, `id_rsa`, `id_ecdsa`. If not found,
use the default path `~/.ssh/id_ed25519` and print a warning message
that asks the user to generate a key pair or update the key path in the
config file.