Add a new ping client message type for `pie-cli ping` command.
Previously, the ping command reported success if the connection and
authentication succeed, or failure otherwise. Now, after establishing
the connection, it sends a ping message to the server and waits for a
pong response. It reports the latency of the ping round trip.

By sending a ping message, we can also use `pie-cli ping` as test cases
when authentication is disabled on one or both sides.