Support re-attaching to a detached inferlet instance.
When an inferlet is re-attached to, its buffered output and future
output will be streamed to the client.

A new `pie-cli attach` command has been added to attach to a running
inferlet instance.

If an inferlet has finished execution, attaching to it will stream its
buffered output and cause the engine to clean up the instance.