Let server send success message before sending streamed output.
There was a race condition where output would be sent to the client
before the success message for instance launch or attach was sent to
the client. In this case, the client would lose the streamed output.

This commit fixes the race condition by sending the success message
before sending the streamed output.